+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added OutMaster rule (thanks James M.)
 - improved generic rules for Qualtrics and Amazon SES (thanks Robert R.)
+- fixed generic tracker rule removing spacers
 
 ## [0.4.13] - 2021-10-22